whatsapp shop/ai/qdrant/requirements.txt
qdrant-client>=1.6.0
python-dotenv>=0.21.0
